+++ packages/expo-maps/CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### 🎉 New features
 
+- [Android] Add programmatic selection support for markers and annotations via `selectedId` prop and `select()` ref method. ([#42189](https://github.com/expo/expo/pull/42189) by [@huextrat](https://github.com/huextrat))
+- [Android] Add `onDeselect` event callback for when a marker is deselected. ([#42189](https://github.com/expo/expo/pull/42189) by [@huextrat](https://github.com/huextrat))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others